Moved logic out of service
